pyproject.toml
[tool.poetry]
name = "smartforecasting"
version = "0.1.0"
description = "SmartForecasting is a cutting-edge service designed to forecast product consumption and sales trends."
authors = ["Abdelmalek Djamaa "]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
numpy = "^2.0.0"
pandas = "^2.2.2"
flask = "^3.0.3"
pydantic = "^2.8.2"
statsmodels = "^0.14.2"
pytest = "^8.3.1"
celery = "^5.4.0"
redis = "^5.0.7"
psycopg2-binary = "^2.9.9"
flasgger = "^0.9.7.1"
icecream = "^2.1.3"
gunicorn = "^22.0.0"
python-fasthtml = "^0.2.1"
flask-cors = "^4.0.1"
fh-matplotlib = "^0.0.3"
python-dateutil = "^2.9.0.post0"
reorder-python-imports = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
